Wine CRUD complete

+++ src/main/java/com/dp3/web/controller/CellarController.java
@@ -1,7 +1,9 @@
 package com.dp3.web.controller;
 
 import com.dp3.dao.CellarRepository;
+import com.dp3.dao.WineRepository;
 import com.dp3.domain.Cellar;
+import com.dp3.domain.Wine;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Sort;
 import org.springframework.http.HttpStatus;
@@ -14,12 +16,14 @@ import javax.validation.Valid;
 import java.util.List;
 
 @RestController
-@RequestMapping("/cellars")
+@RequestMapping("/cellar")
 @CrossOrigin("*")
 public class CellarController {
 
     @Autowired
     private CellarRepository cellarRepository;
+    @Autowired
+    private WineRepository wineRepository;
 
     @GetMapping("/all")
     public List<Cellar> getAllCellars(){
@@ -27,15 +31,15 @@ public class CellarController {
         return cellarRepository.findAll(sortByCellarNameDesc);
     }
 
-    @PostMapping("/createCellar")
-    public ModelAndView createCellar(Model model, @Valid Cellar cellar){
+    @PostMapping("/create")
+    public ModelAndView createCellar(Model model, @Valid @RequestBody  Cellar cellar){
         cellarRepository.save(cellar);
         model.addAttribute("cellars", cellarRepository.findAll());
         return new ModelAndView("redirect:/stock/all");
     }
 
-    @GetMapping(value = "/cellar/{id}")
-    public ResponseEntity<Cellar> getCellarByCellarName(@PathVariable("id") String id){
+    @GetMapping(value = "/{id}")
+    public ResponseEntity<Cellar> getCellarByCellarName(@PathVariable("id") Integer id){
         Cellar cellar = cellarRepository.findOne(id);
         if(cellar == null){
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
@@ -44,23 +48,36 @@ public class CellarController {
         }
     }
 
-    @PutMapping(value = "/cellar/{id}")
-    public ResponseEntity<Cellar> updateCellar(@PathVariable("id") String id,
+    @PutMapping(value = "/{id}")
+    public ResponseEntity<Cellar> updateCellar(@PathVariable("id") Integer id,
                                                @Valid @RequestBody Cellar cellar){
         Cellar cellarData = cellarRepository.findOne(id);
         if(cellarData == null){
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         }
-        cellarData.setCellarName(cellar.getCellarName());
-        cellarData.setDescription(cellar.getDescription());
-        cellarData.setLocation(cellar.getLocation());
+        if (cellar.getCellarName()!=null){
+            cellarData.setCellarName(cellar.getCellarName());
+        }
+        if (cellar.getDescription()!=null){
+            cellarData.setDescription(cellar.getDescription());
+        }
+        if (cellar.getLocation()!=null) {
+            cellarData.setLocation(cellar.getLocation());
+        }
         Cellar updateCellar = cellarRepository.save(cellarData);
         return new ResponseEntity<>(updateCellar,HttpStatus.OK);
     }
 
-    @DeleteMapping(value = "/cellar/{id}")
-    public void deleteCellar(@PathVariable("id") String id){
-        cellarRepository.delete(id);
+    @DeleteMapping(value = "/{id}")
+    public ResponseEntity deleteCellar(@PathVariable("id") Integer id){
+        Cellar cellarToDelete = cellarRepository.findOne(id);
+        List<Wine> wines = wineRepository.findByCellar(cellarToDelete);
+        if (wines.isEmpty()){
+            cellarRepository.delete(id);
+            return new ResponseEntity(HttpStatus.OK);
+        } else {
+            return new ResponseEntity(HttpStatus.METHOD_NOT_ALLOWED);
+        }
     }
 
 }
